--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,9 +1,12 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
 
 import (
+	"archive/zip"
 	"bytes"
+	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -36,4 +39,53 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+}
+
+func TestGetDomainStat2(t *testing.T) {
+	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov.com","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@browsecat.BroWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.gov","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@com.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+	t.Run("find 'com'", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{
+			"browsecat.com":           1,
+			"browsecat.browsecat.com": 1,
+			"browsedrive.gov.com":     1,
+			"com.com":                 1,
+		}, result)
+	})
+
+	t.Run("find 'gov'", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "gov")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{"teklist.gov": 1}, result)
+	})
+
+	t.Run("find broken domain", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "r'(?:^|\\s)([\\w\\d]+)=((?:[^\\\\\\s=]+|\\\\.)+)(?=\\s|$)'")
+		require.Truef(t, errors.Is(err, ErrInvalidDomain), "actual err - %v", err)
+		require.Equal(t, DomainStat{}, result)
+	})
+
+	brokenData := `BROKEN FORMAT`
+	t.Run("broken format usecase", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(brokenData), "com")
+		require.Truef(t, errors.Is(err, ErrInvalidFileData), "actual err - %v", err)
+		require.Equal(t, DomainStat{}, result)
+	})
+}
+
+func BenchmarkGetDomainStat(b *testing.B) {
+	r, err := zip.OpenReader("testdata/users.dat.zip")
+	require.NoError(b, err)
+
+	data, err := r.File[0].Open()
+	require.NoError(b, err)
+
+	for i := 0; i < b.N; i++ {
+		GetDomainStat(data, "com")
+	}
 }