@xsicx 2022-07-17
**old**
```
goos: darwin
goarch: amd64
pkg: github.com/xsicx/otus-go/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetDomainStat-12    	    3121	   3967173 ns/op	  136430 B/op	    1224 allocs/op
PASS
ok  	github.com/xsicx/otus-go/hw10_program_optimization	24.332s
```
**new**
```
goos: darwin
goarch: amd64
pkg: github.com/xsicx/otus-go/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetDomainStat-12    	 4879242	      2106 ns/op	    5525 B/op	      21 allocs/op
PASS
ok  	github.com/xsicx/otus-go/hw10_program_optimization	11.825s
```
**benchstat**
```
name              old time/op    new time/op    delta
GetDomainStat-12    3.97ms ± 0%    0.00ms ± 0%  -99.95%

name              old alloc/op   new alloc/op   delta
GetDomainStat-12     136kB ± 0%       6kB ± 0%  -95.95%

name              old allocs/op  new allocs/op  delta
GetDomainStat-12     1.22k ± 0%     0.02k ± 0%  -98.28%
```